+++ src/App/images/index.ts
@@ -1,0 +1,1 @@
+export {default as logo} from './logo.svg';
