+++ src/App/index.ts
@@ -1,0 +1,4 @@
+import App from './App';
+
+export {default as theme} from './theme';
+export default App;
