+++ src/server/index.ts
@@ -1,0 +1,3 @@
+import * as serviceWorker from './serviceWorker';
+
+export {serviceWorker};
